+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [Upcoming]
+## [1.0.1] - 2021-12-07
+
 * `rasterise` function now can work with whole ggplot objects
 
 ## [1.0.0] - 2021-11-15